requirements.txt
django-ipware==2.1.0
xxhash==2.0.0
